+++ app/Filament/Resources/Announcements/AnnouncementResource.php
@@ -15,7 +15,6 @@ use Filament\Resources\Resource;
 use Filament\Schemas\Schema;
 use Filament\Support\Icons\Heroicon;
 use Filament\Tables\Table;
-use Illuminate\Database\Eloquent\Builder;
 use UnitEnum;
 
 class AnnouncementResource extends Resource
@@ -34,12 +33,6 @@ class AnnouncementResource extends Resource
     public static function table(Table $table): Table
     {
         return AnnouncementsTable::configure($table);
-    }
-
-    public static function getEloquentQuery(): Builder
-    {
-        return parent::getEloquentQuery()
-            ->withCount(['views as dismissed_count' => fn (Builder $query) => $query->whereNotNull('dismissed_at')]);
     }
 
     public static function getPages(): array